# docs/api/openapi-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Millet ERP API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/catalogos/proveedores": {
      "get": {
        "tags": [
          "Catalogos"
        ],
        "summary": "Listar proveedores cross-empresa",
        "description": "Lista paginada de proveedores en `compartido.proveedores`. Filtros opcionales: `estatus` (Activo/Inactivo/EnRevisión), `clave` (substring match). Paginación offset-based con tope de `limit=200`. Permiso: `compartido.catalogos.leer`.",
        "operationId": "ListarProveedores",
        "parameters": [
          {
            "name": "estatus",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/NullableOfEstatusCatalogo"
            }
          },
          {
            "name": "clave",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedCatalogoResponseOfProveedorListItem"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Catalogos"
        ],
        "summary": "Crear proveedor (B.5)",
        "description": "Alta de proveedor. UNIQUE(clave) → 422 `PROVEEDOR_CLAVE_DUPLICADA` si choca. Si llega `monedaPreferidaId`, validación cross-table contra `compartido.monedas` (404 si no existe). Header `Idempotency-Key` obligatorio.",
        "operationId": "CrearProveedor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearProveedorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrearProveedorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalogos/proveedores/{id}": {
      "get": {
        "tags": [
          "Catalogos"
        ],
        "summary": "Detalle de proveedor por id",
        "description": "Devuelve cabecera completa del proveedor (incluye `email`, `telefono`, `claveLegacy` que el list no expone). 404 si el id no existe.",
        "operationId": "ObtenerProveedorPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProveedorDetalle"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Catalogos"
        ],
        "summary": "Editar proveedor (PATCH parcial, B.5)",
        "description": "PATCH parcial sobre los campos editables. Convención: nullable `null` = no tocar; flag `limpiarX = true` = setear nullable a null. Inmutables: `id`, `clave`, `claveLegacy`. Header `Idempotency-Key` obligatorio.",
        "operationId": "ActualizarProveedor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarProveedorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Catalogos"
        ],
        "summary": "Desactivar proveedor (soft delete, B.5)",
        "description": "Soft delete: setea `Estatus = Inactivo`. Idempotente: si ya está Inactivo, retorna 204 sin cambios. Las RQs históricas siguen funcionando con el id; las nuevas RQs se bloquean por la validación cross-table de F7-PR1. Para reactivar, usar PATCH (cuando se necesite).",
        "operationId": "DesactivarProveedor",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActualizarProveedorRequest": {
        "required": [
          "razonSocial",
          "nombreComercial",
          "rfc",
          "tipoPersona",
          "condicionesPagoDias",
          "monedaPreferidaId",
          "email",
          "telefono",
          "limpiarNombreComercial",
          "limpiarCondicionesPago",
          "limpiarMonedaPreferida",
          "limpiarEmail",
          "limpiarTelefono"
        ],
        "type": "object",
        "properties": {
          "razonSocial": {
            "type": "string",
            "nullable": true
          },
          "nombreComercial": {
            "type": "string",
            "nullable": true
          },
          "rfc": {
            "type": "string",
            "nullable": true
          },
          "tipoPersona": {
            "$ref": "#/components/schemas/NullableOfTipoPersonaProveedor"
          },
          "condicionesPagoDias": {
            "type": "integer",
            "format": "int16",
            "nullable": true
          },
          "monedaPreferidaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "limpiarNombreComercial": {
            "type": "boolean",
            "nullable": true
          },
          "limpiarCondicionesPago": {
            "type": "boolean",
            "nullable": true
          },
          "limpiarMonedaPreferida": {
            "type": "boolean",
            "nullable": true
          },
          "limpiarEmail": {
            "type": "boolean",
            "nullable": true
          },
          "limpiarTelefono": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
...
      "CrearProveedorRequest": {
        "required": [
          "clave",
          "razonSocial",
          "rfc",
          "tipoPersona",
          "nombreComercial",
          "condicionesPagoDias",
          "monedaPreferidaId",
          "email",
          "telefono"
        ],
        "type": "object",
        "properties": {
          "clave": {
            "type": "string"
          },
          "razonSocial": {
            "type": "string"
          },
          "rfc": {
            "type": "string"
          },
          "tipoPersona": {
            "$ref": "#/components/schemas/TipoPersonaProveedor"
          },
          "nombreComercial": {
            "type": "string",
            "nullable": true
          },
          "condicionesPagoDias": {
            "type": "integer",
            "format": "int16",
            "nullable": true
          },
          "monedaPreferidaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CrearProveedorResponse": {
        "required": [
          "id",
          "clave"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clave": {
            "type": "string"
          }
        }
      },
...
      "EstatusCatalogo": {
        "type": "integer"
      },
...
      "NullableOfEstatusCatalogo": {
        "type": "integer"
      },
...
      "NullableOfTipoPersonaProveedor": {
        "type": "integer",
        "nullable": true
      },
...
      "PagedCatalogoResponseOfProveedorListItem": {
        "required": [
          "items",
          "offset",
          "limit",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProveedorListItem"
            }
          },
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "limit": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ProveedorDetalle": {
        "required": [
          "id",
          "clave",
          "claveLegacy",
          "razonSocial",
          "nombreComercial",
          "rfc",
          "tipoPersona",
          "condicionesPagoDias",
          "monedaPreferidaId",
          "email",
          "telefono",
          "estatus"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clave": {
            "type": "string"
          },
          "claveLegacy": {
            "type": "string",
            "nullable": true
          },
          "razonSocial": {
            "type": "string"
          },
          "nombreComercial": {
            "type": "string",
            "nullable": true
          },
          "rfc": {
            "type": "string"
          },
          "tipoPersona": {
            "$ref": "#/components/schemas/TipoPersonaProveedor"
          },
          "condicionesPagoDias": {
            "type": "integer",
            "format": "int16",
            "nullable": true
          },
          "monedaPreferidaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "estatus": {
            "$ref": "#/components/schemas/EstatusCatalogo"
          }
        }
      },
      "ProveedorListItem": {
        "required": [
          "id",
          "clave",
          "razonSocial",
          "nombreComercial",
          "rfc",
          "tipoPersona",
          "condicionesPagoDias",
          "monedaPreferidaId",
          "estatus"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clave": {
            "type": "string"
          },
          "razonSocial": {
            "type": "string"
          },
          "nombreComercial": {
            "type": "string",
            "nullable": true
          },
          "rfc": {
            "type": "string"
          },
          "tipoPersona": {
            "$ref": "#/components/schemas/TipoPersonaProveedor"
          },
          "condicionesPagoDias": {
            "type": "integer",
            "format": "int16",
            "nullable": true
          },
          "monedaPreferidaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "estatus": {
            "$ref": "#/components/schemas/EstatusCatalogo"
          }
        }
      },
...
      "TipoPersonaProveedor": {
        "type": "integer"
      },
...
    }
  },
...
}